symlinks don't work; using a custom file to linke dependencies...
