+++ exa/static/nbextensions/__paths__
@@ -1,0 +1,3 @@
+'js/lib/three.min.js'
+'js/lib/TrackballControls.js'
+'js/tester.js'
