--- exa/static/nbextensions/apps/app3d.js
@@ -1,1 +1,0 @@
-../../js/apps/app3d.js
