--- exa/static/nbextensions/lib/TrackballControls.js
@@ -1,1 +1,0 @@
-../../js/lib/TrackballControls.js
